Added Firewire 800 Support to ICL's DCGrabber interface
And suprise suprise: It is working!

I checked this new feature with a Point Grey
Flea-2 camera in YUV-422 Mode 1024x768 @ 30 Hz

This mode needs a transfer rate of about 500MBits,
which is not available using the legacy mode of
the IEEE-1394 Interface (I guess).

To use this new feature, several functions have been
added at several locations:
- iclDC.h void set_iso_speed(...)
  is a global utility function which cam be
  called with a camera and an integer valued
  iso speed in MBits. It sets up both, iso speed
  itself and an propropiate operation mode
- A DCGrabber can now be instantiated now with an additional
  isoSpeedInMBits value
- GenericGrabbers can be instantiated with a 'dc800' string,
  which causes internal dc grabbers a instantiated with an
  iso speed of 800 MBits
- the camcfg tool can now be started with an optinal flag
  '-800' which does what it sounds like