added a toggle button implementation.
The former impl. turned out not to work correctly -- in some
situations, the toggled text was shown in un-toggle state and
vice versa. The new ToggleButton class wraps a QPushButton 
very shallowly using inheritance. Internally it connects the
QPushButtons checked-signal to a private slot, where the button
text is changed. Additionally it provides the ability of 
wrapping a given bool-pointer, which's value is immediately 
changed each time the toggle-state of the button changes.

The "togglebutton"-GUI component uses now an instance of
this new ToggleButton class, however the ButtonHandle 
registered at the GUI's datastore does still wrap an instance
of QPushButton. This was trivial, as ToggleButton inherits
QPushButton in public manner.