Added a new internal feature to about one half the the ICLQuick-functions
In the current implementation, the adapted functions are internally 
re-using temporary images if available (and currently not referenced 
somewhere else). 
To avoid some implementational overhead, I added a hidden function called
get_temp_image, which handles an internal buffer of temporary images. Each
time, an ICLQuick function needs a temporary image, the list of already
created temporary images is searched for a temporary image, which is currently
not referenced somewhere else, and which is has compatible parameters (size
and channel count) If one of these images is available, it is used instead of
creating a new image. If no image is compatible, get_temp_image searches for
a currently not-shared image within the static list. Only if no image is 
found by this means too, a new temporary image is put into the map.

Additionally I added a simple example, which shows performance comparison
using a lot of queued ICLQuick functions. The optimization enhanced performance
by about 100% (doubled framerate)