added yet another feature to the ConfigFile
now it's possible (in addtion to the range=[min,max] restriction
of int- and float-entries) to restrict float entries with a
values="{a,b,c}" etc attribute within the data tag. So e.g.

<data id="grabber-type" type="string" values="[pwc,dc,file,unicap]"\>dc\</data\>

makes ConfigFileGUI attached to the ConfigFile create a combo-box
with all value alternative of the values-list (please ensure you have
used some enclosing brackets around the comma-separated value list)