Tryed to add the doxylink extension: Finally i was able to do so, but it seems to be more complex than i hoped:
First of all, the extension needs to be moved to the build-directory, or
we must consider an in-source build with cmake (dunno)

The extension needs the doxygen-tag-file, which i was able to build,
however, the extension seems to have problems with build-errors within
the doxygen documentation.

Right now, I used the alldoc documentation, which will be the only
one we'll keep in the future, perhaps, we could re-try this step by
making alldoc only contain the ICLUtils parts first.
Sooner or later we will have to fix the doxygen generation completely!