created a link between the scene and the ICLWidget it is drawn in
Now, special MouseHandlers can implement a link and an unlink method
that can be used to affect the ICLWidget instance when the
MouseHandler is install to it.  The Scene's SceneMouseHandler uses
this mechanism to add an extra OSD button that shows a scene-specific
extra GUI. 

So far, only a "proof-of-concept"-GUI is created and shown. Soon
i plan provide the following features in this GUI:

* enable disable lights
* enable disable lighting
* show/hide lights
* show the scene coordinate frame
* show coordinate frames for each object 

For these features, it is worth thinking about
making the scene a Configurable instance.