Generate Package data first; Fix ICL_FIND_PACKAGE for IPP 9+
this way cmake can configure the package data before all dependencies are installed.
Since IPP 9+ does not include ippm any more the detection step has to be done in two steps: first look for the OLD/legacy version. if this is not found search for a newer version (without ippm)